requirements.txt
asgiref==3.5.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Django==4.0.3
django-debug-toolbar==3.2.4
gunicorn==20.1.0
idna==3.3
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
psycopg2==2.9.3
python-dotenv==0.19.2
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.9
